Move flag serialization over to a manager
Custom flag formats are now handled by a manager, instead of being
static and therefore thread-shared.
